Create date calendar
Certainly! Here's an example of how you can write the `getDay` method to retrieve the day of the week for a given date using the `Calendar` class in Java 7:

```java
import java.util.*;

public class DayOfWeek {
    public static void main(String[] args) {
        Scanner scanner = new Scanner(System.in);
        System.out.println("Enter a date (yyyy-mm-dd): ");
        String inputDate = scanner.nextLine();

        String dayOfWeek = getDay(inputDate);
        System.out.println("Day of the week: " + dayOfWeek);
    }

    public static String getDay(String inputDate) {
        try {
            // Split the input date string to retrieve year, month, and day
            String[] dateParts = inputDate.split("-");
            int year = Integer.parseInt(dateParts[0]);
            int month = Integer.parseInt(dateParts[1]);
            int day = Integer.parseInt(dateParts[2]);

            // Create a Calendar instance and set the given date
            Calendar calendar = Calendar.getInstance();
            calendar.set(year, month - 1, day);

            // Get the day of the week as an integer (1 = Sunday, 2 = Monday, ..., 7 = Saturday)
            int dayOfWeekInt = calendar.get(Calendar.DAY_OF_WEEK);

            // Convert the day of the week integer to the corresponding day name
            String[] daysOfWeek = {"Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"};
            String dayOfWeek = daysOfWeek[dayOfWeekInt - 1];

            return dayOfWeek;
        } catch (Exception e) {
            System.out.println("Invalid date format. Please enter a valid date in yyyy-mm-dd format.");
            return null;
        }
    }
}
```

In this code, we prompt the user to enter a date in the format `yyyy-mm-dd`. Then we call the `getDay` method, passing the input date as an argument. The `getDay` method splits the input date into year, month, and day, and creates a `Calendar` instance. We set the year, month, and day using the `set` method of the `Calendar` class. Then we retrieve the day of the week as an integer using `Calendar.DAY_OF_WEEK` and convert it to the corresponding day name using an array of day names. Finally, we return the day of the week as a string.

Note: This code assumes that the input date is in the correct format (`yyyy-mm-dd`). It also handles exceptions and displays an error message if the date format is invalid.